+++ src/renderers/React.js
@@ -29,8 +29,9 @@ export default class ReactRenderer extends Renderer {
 
   beforeParse() {
     this.imports.push('import PropTypes from \'prop-types\';');
+
     if (this.options.stripPropTypes) {
-      this.header.push('const __productionShape__ = () => {}');
+      this.header.push('const PropTypePolyfill = () => {};');
     }
   }
 
@@ -38,7 +39,7 @@ export default class ReactRenderer extends Renderer {
     const shape = this.formatObject(this.renderObjectProps(attributes, 1), 0);
 
     if (this.options.stripPropTypes) {
-      return `export const ${setName} = process.env.NODE_ENV === 'production' ? __productionShape__ : PropTypes.shape(${shape});`;
+      return `export const ${setName} = (process.env.NODE_ENV === 'production') ? PropTypePolyfill : PropTypes.shape(${shape});`;
     }
 
     return `export const ${setName} = PropTypes.shape(${shape});`;